Split apart MemoryBoard
